moved Config from utils to config package
