+++ pom.xml
@@ -106,6 +106,6 @@
             <artifactId>commons-lang3</artifactId>
             <version>3.1</version>
         </dependency>
-        
+
     </dependencies>
 </project>
